// React-with-types/courseApp/src/types.ts
export interface HeaderProps {
    title: string;
}

export interface PartProps {
    part: CoursePart;
}

export interface ContentList {
    contents: CoursePart[]
}

interface CoursePartBase {
    name: string;
    exerciseCount: number;
}

interface CoursePartBasic extends CoursePartBase {
    description: string;
    kind: "basic"
}

interface CoursePartGroup extends CoursePartBase {
    groupProjectCount: number;
    kind: "group"
}

interface CoursePartBackground extends CoursePartBase {
    description: string;
    backgroundMaterial: string;
    kind: "background"
}

interface CoursePartSpecial extends CoursePartBase {
    description: string;
    requirements: string[];
    kind: "special"
}

export type CoursePart =
    CoursePartBasic
    | CoursePartGroup
    | CoursePartBackground
    | CoursePartSpecial;